// client/api/models/CreateWorkerAssignmentRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Smart Power Electric API
 * Smart Power Electric API Documentation
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateWorkerAssignmentRequest
 */
export interface CreateWorkerAssignmentRequest {
    /**
     * Name of the workerAssignment
     * @type {string}
     * @memberof CreateWorkerAssignmentRequest
     */
    workerId: string | null;
    /**
     * Project ID of the workerAssignment
     * @type {string}
     * @memberof CreateWorkerAssignmentRequest
     */
    projectId: string | null;
    /**
     * Stage ID of the workerAssignment
     * @type {string}
     * @memberof CreateWorkerAssignmentRequest
     */
    stageId: string | null;
}

/**
 * Check if a given object implements the CreateWorkerAssignmentRequest interface.
 */
export function instanceOfCreateWorkerAssignmentRequest(value: object): value is CreateWorkerAssignmentRequest {
    if (!('workerId' in value) || value['workerId'] === undefined) return false;
    if (!('projectId' in value) || value['projectId'] === undefined) return false;
    if (!('stageId' in value) || value['stageId'] === undefined) return false;
    return true;
}

export function CreateWorkerAssignmentRequestFromJSON(json: any): CreateWorkerAssignmentRequest {
    return CreateWorkerAssignmentRequestFromJSONTyped(json, false);
}

export function CreateWorkerAssignmentRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateWorkerAssignmentRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'workerId': json['workerId'],
        'projectId': json['projectId'],
        'stageId': json['stageId'],
    };
}

  export function CreateWorkerAssignmentRequestToJSON(json: any): CreateWorkerAssignmentRequest {
      return CreateWorkerAssignmentRequestToJSONTyped(json, false);
  }

  export function CreateWorkerAssignmentRequestToJSONTyped(value?: CreateWorkerAssignmentRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'workerId': value['workerId'],
        'projectId': value['projectId'],
        'stageId': value['stageId'],
    };
}