// client/api/model/updateWorkerAssignmentRequest.ts
/**
 * Smart Power Electric API
 * Smart Power Electric API Documentation
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class UpdateWorkerAssignmentRequest {
    /**
    * Name of the workerAssignment
    */
    'workerId': string | null;
    /**
    * Project ID of the workerAssignment
    */
    'projectId': string | null;
    /**
    * Stage ID of the workerAssignment
    */
    'stageId': string | null;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "workerId",
            "baseName": "workerId",
            "type": "string"
        },
        {
            "name": "projectId",
            "baseName": "projectId",
            "type": "string"
        },
        {
            "name": "stageId",
            "baseName": "stageId",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return UpdateWorkerAssignmentRequest.attributeTypeMap;
    }
}